implement camera transform

+++ src/multiplication.hpp
@@ -6,8 +6,8 @@
 class multiplication {
 
 public:
-  static std::vector<std::vector<int> > result(const std::vector<std::vector<int> >& A,
-                                 const std::vector<std::vector<int> >& B);
+  static std::vector<std::vector<double> > result(const std::vector<std::vector<double> >& A,
+                                 const std::vector<std::vector<double> >& B);
 
 };
